Adds basic structure for parsing markdown
This process is lossy, but may be sufficient for this project. A full archive
is still nice to have (for me), but is likely not needed for most people
and certainly not for the LLM. This will save a ton of tokens and simplify
the parsing method.

This is still limited to websites that will allow a basic curl. I can
consider including a pluggable option for a browser eventually, but at
most I can provide a reasonable UserAgent to "fake" the basics.

JS sites will not work until a headless browser is plugged in, which
increases the cost significantly. This would be okay for my use, but not for
general user (maybe that's another monetization option).